swatch_pipeline_ml/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.1
setuptools==80.9.0
wheel==0.45.1
